fixed data types

+++ array_2d.c
@@ -1,13 +1,13 @@
 #include "array_2d.h"
 
-void destroy_array_2d(size_t ***ptr, const size_t height)
+void destroy_array_2d(signed long long int ***ptr, const size_t height)
 {
     if (ptr == NULL || height == 0)
     {
         return;
     }
 
-    size_t **temp = *ptr;
+    signed long long int **temp = *ptr;
     for (size_t i = 0; i < height; i++)
     {
         free(*(temp + i));
@@ -17,14 +17,14 @@ void destroy_array_2d(size_t ***ptr, const size_t height)
     *ptr = NULL;
 }
 
-int create_array_2d(size_t ***ptr, const size_t width, const size_t height)
+int create_array_2d(signed long long int ***ptr, const size_t width, const size_t height)
 {
     if(width == 0 || height == 0 || ptr == NULL)
     {
         return 1;
     }
 
-    size_t **array = malloc(sizeof(size_t *) * height);
+    signed long long int **array = malloc(sizeof(signed long long int *) * height);
     if(array == NULL)
     {
         return 2;
@@ -32,7 +32,7 @@ int create_array_2d(size_t ***ptr, const size_t width, const size_t height)
 
     for(size_t i = 0; i < height; i++)
     {
-        *(array + i) = malloc(sizeof(size_t) * width);
+        *(array + i) = malloc(sizeof(signed long long int) * width);
         if(*(array + i) == NULL)
         {
             destroy_array_2d(&array, i);
@@ -45,22 +45,22 @@ int create_array_2d(size_t ***ptr, const size_t width, const size_t height)
     return 0;
 }
 
-int create_array_2d_init(size_t ***ptr, const size_t width, const size_t height, const size_t value)
+int create_array_2d_init(signed long long int ***ptr, const size_t width, const size_t height, const signed long long int value)
 {
     if(width == 0 || height == 0 || ptr == NULL)
     {
         return 1;
     }
 
-    size_t **array = malloc(sizeof(size_t *) * height);
+    signed long long int **array = malloc(sizeof(signed long long int *) * height);
     if(array == NULL)
     {
         return 2;
     }
 
-    for(size_t i = 0; i < height; i++)
+    for(signed long long int i = 0; i < height; i++)
     {
-        *(array + i) = malloc(sizeof(size_t) * width);
+        *(array + i) = malloc(sizeof(signed long long int) * width);
         if(*(array + i) == NULL)
         {
             destroy_array_2d(&array, i);
@@ -81,14 +81,14 @@ int create_array_2d_init(size_t ***ptr, const size_t width, const size_t height,
     return 0;
 }
 
-int create_array_2d(size_t ***ptr, const size_t size)
+int create_array_2d(signed long long int ***ptr, const size_t size)
 {
     if(size == 0 || ptr == NULL)
     {
         return 1;
     }
 
-    size_t **array = malloc(sizeof(size_t *) * size);
+    signed long long int **array = malloc(sizeof(signed long long int *) * size);
     if (array == NULL)
     {
         return 2;
@@ -96,7 +96,7 @@ int create_array_2d(size_t ***ptr, const size_t size)
 
     for(size_t i = 0; i < size; i++)
     {
-        *(array + i) = malloc(sizeof(size_t) * size);
+        *(array + i) = malloc(sizeof(signed long long int) * size);
         if (*(array + i) == NULL)
         {
             destroy_array_2d(&array, i);
@@ -109,14 +109,14 @@ int create_array_2d(size_t ***ptr, const size_t size)
     return 0;
 }
 
-int create_array_2d_init(size_t ***ptr, const size_t size, const size_t value)
+int create_array_2d_init(signed long long int ***ptr, const size_t size, const signed long long int value)
 {
     if(size == 0 || ptr == NULL)
     {
         return 1;
     }
 
-    size_t **array = malloc(sizeof(size_t *) * size);
+    signed long long int **array = malloc(sizeof(signed long long int *) * size);
     if (array == NULL)
     {
         return 2;
@@ -124,7 +124,7 @@ int create_array_2d_init(size_t ***ptr, const size_t size, const size_t value)
 
     for(size_t i = 0; i < size; i++)
     {
-        *(array + i) = malloc(sizeof(size_t) * size);
+        *(array + i) = malloc(sizeof(signed long long int) * size);
         if (*(array + i) == NULL)
         {
             destroy_array_2d(&array, i);
@@ -145,7 +145,7 @@ int create_array_2d_init(size_t ***ptr, const size_t size, const size_t value)
     return 0;
 }
 
-void display_array_2d(size_t **ptr, const size_t width, const size_t height)
+void display_array_2d(signed long long int **ptr, const size_t width, const size_t height)
 {
     if (ptr == NULL || width == 0 || height == 0)
     {
